+++ randovania/resolver/resolver.py
@@ -177,6 +177,7 @@ async def resolve(configuration: BaseConfiguration,
 
     game = default_database.game_description_for(configuration.game).make_mutable_copy()
     derived_nodes.create_derived_nodes(game)
+    derived_nodes.remove_inactive_layers(game, configuration.active_layers())
     bootstrap = game.game.generator.bootstrap
 
     game.resource_database = bootstrap.patch_resource_database(game.resource_database, configuration)